requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
black==25.1.0
bleach==6.1.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.2.0
cryptography==42.0.1
cssselect2==0.7.0
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
distlib==0.3.7
distro==1.9.0
Django==5.0
django-crispy-forms==1.12.0
dnspython==2.7.0
duckdb==1.1.3
duckduckgo_search==6.4.2
ecdsa==0.19.1
effdet==0.4.1
email_validator==2.2.0
emoji==2.14.0
eval_type_backport==0.2.0
executing==2.0.1
faiss-cpu==1.8.0.post1
fastapi==0.115.13
fastjsonschema==2.19.1
feedparser==6.0.11
filelock==3.12.4
filetype==1.2.0
Flask==2.1.2
Flask-Cors==4.0.0
flatbuffers==24.3.25
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
future==0.18.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-vision==3.9.0
googleapis-common-protos==1.66.0
gpt_index==0.2.6
greenlet==3.0.3
groq==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==20.0.4
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.6
iniconfig==2.1.0
InstructorEmbedding==1.0.1
iopath==0.1.10
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.2
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.5.0
joblib==1.4.2
json5==0.9.24
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-groq==0.2.2
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.4
langdetect==1.0.9
langsmith==0.2.4
layoutparser==0.3.4
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.22.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
narwhals==1.6.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.1.2
notebook_shim==0.2.4
numexpr==2.10.2
numpy==1.26.3
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
openai==0.27.4
opencv-python==4.10.0.84
orjson==3.10.7
oscrypto==1.3.0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
passlib==1.7.4
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
php==1.2.1
pi_heif==0.21.0
pikepdf==9.4.2
pillow==10.2.0
pipenv==2023.10.24
platformdirs==3.11.0
pluggy==1.6.0
portalocker==3.0.0
primp==0.9.1
prometheus_client==0.20.0
prompt-toolkit==3.0.39
proto-plus==1.25.0
protobuf==5.28.0
psutil==5.9.6
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.7.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.16.1
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyMuPDF==1.25.1
pyparsing==3.1.1
pypdf==4.0.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pypng==0.20220715.0
pyreadline3==3.5.4
pytest==8.4.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2024.10.22
python-jose==3.5.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.1
qrcode==7.4.2
qtconsole==5.5.1
QtPy==2.4.1
RapidFuzz==3.11.0
referencing==0.34.0
regex==2024.7.24
reportlab==4.0.9
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rpds-py==0.18.0
rsa==4.9
ruff==0.12.0
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.12.0
seaborn==0.11.1
Send2Trash==1.8.2
sentence-transformers==3.0.1
setuptools==68.2.2
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.46.2
streamlit==1.38.0
svglib==1.5.1
sympy==1.13.1
tenacity==8.5.0
terminado==0.18.1
textblob==0.18.0.post0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.12
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.19.1
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.3.3
tqdm==4.66.5
traitlets==5.13.0
transformers==4.44.2
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
unstructured==0.16.11
unstructured-client==0.28.1
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
uri-template==1.3.0
uritools==4.0.2
urllib3==1.26.20
uvicorn==0.34.3
validators==0.28.1
virtualenv==20.24.5
virtualenvwrapper-win==1.2.7
watchdog==4.0.2
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.2.2
widgetsnbextension==4.0.10
wikipedia==1.4.0
wrapt==1.17.0
xhtml2pdf==0.2.14
yarl==1.9.7
youtube-transcript-api==0.6.3
